Make `PackageURLBuilder::getQualifiers` return `Map`
The return type of the method `PackageURLBuilder::getQualifiers` should
be an interface, such as `Map`, rather than the implementation `TreeMap`.
This aligns with the return type of the method
`PackageURL::getQualifiers`.

Also, make the return types of this method match the return types of
`PackageURL::getQualifiers`.

Note that this a potentially breaking API change.