Fix URI creation
Remove the `scheme` field and replace it with a `SCHEME` constant as
its impossible to create a `PackageURL` with a different `scheme`.
This should not change the public API.

It was previously the case that parsing a purl could produce different
`URI` representations internally based on the number of slashes present
after the scheme colon in the input. For example, a purl with the
canonical form `pkg:generic/name` when input starting with `pkg:`,
`pkg:/`, and `pkg://` will produce three different `URI`
representations. By normalizing the input to contain a single `'/'`, as
in `pkg:/generic/name`, we obtain a valid hierarchical `URI` with all
five standard components `[scheme:][//authority][path][?query][#fragment]`
properly parsed. For the `URI` to be a valid purl, `scheme:` must be
`"pkg:"`, `authority` must be `null`, and `?query` and `#fragment` are
optional. Then, we only have to parse the `path` containing the
`type/namespace/name@version` portion of the purl.